incomplete rows
In order to work around browsers' different rounding behaviors,
Foundation will float the last column in a row to the right so the edge
aligns. If your row doesn't have a count that adds up to 12 columns, you
can tag the last column with a class of .end in order to override that
behavior.